Use scm-publish:publish-scm instead.
This is to handle some of the maven peculiarities.

This relies on the fact that the JOB_NAME (the dirname part)
is the same as the repo name and that it is the repo name
that is used when maven calculates where to store the result.
After maven has generated the site in the place, it is moved
to the right place before publishing.