Publish test result from first run of tests.
After the change to jacoco, the tests are explicitly run when creating
site. In the case they fail, it is better to have Jenkins publish
the result then because the second run of the tests, when doing install
will then never be performed.

Also, removed the second run of the tests. They are probably still run
implicitly from install and deploy.